# Make Format actually faster