Format is now 1% ns faster on average and >30% worse in memory consumption than fmt. Note that I didn't change the fmt tests at all so this means that there the benchmarks are +- 1% accurate. Also, FormatComplex7Arg seems to jump +-17% speed at random.
```
name                  old time/op    new time/op    delta
Format4Arg-24            520ns ± 1%     510ns ± 1%  -1.87%  (p=0.000 n=10+10)
Fmt4Arg-24               523ns ± 1%     518ns ± 1%  -0.83%  (p=0.000 n=10+9)
Format6Arg-24            554ns ± 1%     545ns ± 1%  -1.59%  (p=0.000 n=10+10)
Fmt6Arg-24               552ns ± 1%     556ns ± 1%  +0.80%  (p=0.000 n=10+10)
FormatComplex7Arg-24     313ns ± 1%     313ns ± 0%    ~     (p=0.644 n=10+9)

name                  old alloc/op   new alloc/op   delta
Format4Arg-24             352B ± 0%      352B ± 0%    ~     (all equal)
Fmt4Arg-24                304B ± 0%      304B ± 0%    ~     (all equal)
Format6Arg-24             416B ± 0%      416B ± 0%    ~     (all equal)
Fmt6Arg-24                320B ± 0%      320B ± 0%    ~     (all equal)
FormatComplex7Arg-24      326B ± 0%      326B ± 0%    ~     (all equal)

name                  old allocs/op  new allocs/op  delta
Format4Arg-24             8.00 ± 0%      8.00 ± 0%    ~     (all equal)
Fmt4Arg-24                5.00 ± 0%      5.00 ± 0%    ~     (all equal)
Format6Arg-24             8.00 ± 0%      8.00 ± 0%    ~     (all equal)
Fmt6Arg-24                5.00 ± 0%      5.00 ± 0%    ~     (all equal)
FormatComplex7Arg-24      3.00 ± 0%      3.00 ± 0%    ~     (all equal)
```